backend/requirements.dev.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Flask==3.1.0
Flask-Login==0.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
